# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Location {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  lat: number;
  lng: number;
  imageUrl: string;
  /** @nullable */
  hint?: string | null;
  createdAt: string;
}

export interface LocationInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  lat: number;
  lng: number;
  /** @minLength 1 */
  imageUrl: string;
  hint?: string;
}

export interface LocationPatch {
  /** @minLength 1 */
  name?: string;
  description?: string;
  lat?: number;
  lng?: number;
  imageUrl?: string;
  hint?: string;
}

export interface LocationStats {
  totalLocations: number;
  totalGames: number;
  avgScore: number;
}

export interface GameInput {
  /**
     * @minimum 1
     * @maximum 10
     */
  rounds?: number;
}

export type GameStatus = typeof GameStatus[keyof typeof GameStatus];


export const GameStatus = {
  in_progress: 'in_progress',
  finished: 'finished',
} as const;

export interface Game {
  id: number;
  rounds: number;
  currentRound: number;
  totalScore: number;
  status: GameStatus;
  /** @nullable */
  currentPhotoUrl: string | null;
  /** @nullable */
  currentLocationName?: string | null;
}

export interface GuessInput {
  lat: number;
  lng: number;
}

export interface GuessResult {
  distanceKm: number;
  score: number;
  correctLat: number;
  correctLng: number;
  correctName: string;
  correctImageUrl?: string;
  totalScore: number;
  roundComplete: boolean;
  gameComplete: boolean;
}

export interface RoundResult {
  round: number;
  locationName: string;
  guessLat: number;
  guessLng: number;
  correctLat: number;
  correctLng: number;
  distanceKm: number;
  score: number;
}

export interface GameResults {
  id: number;
  totalScore: number;
  rounds: number;
  roundResults: RoundResult[];
}