# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatus } from './gameStatus';

export interface Game {
  id: number;
  rounds: number;
  currentRound: number;
  totalScore: number;
  status: GameStatus;
  /** @nullable */
  currentPhotoUrl: string | null;
  /** @nullable */
  currentLocationName?: string | null;
}