# lib/api-zod/src/generated/types/gameInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */

export interface GameInput {
  /**
     * @minimum 1
     * @maximum 10
     */
  rounds?: number;
}