# lib/api-zod/src/generated/types/gameResults.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */
import type { RoundResult } from './roundResult';

export interface GameResults {
  id: number;
  totalScore: number;
  rounds: number;
  roundResults: RoundResult[];
}