# lib/api-zod/src/generated/types/guessResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */

export interface GuessResult {
  distanceKm: number;
  score: number;
  correctLat: number;
  correctLng: number;
  correctName: string;
  correctImageUrl?: string;
  totalScore: number;
  roundComplete: boolean;
  gameComplete: boolean;
}