# lib/api-zod/src/generated/types/location.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */

export interface Location {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  lat: number;
  lng: number;
  imageUrl: string;
  /** @nullable */
  hint?: string | null;
  createdAt: string;
}