# lib/api-zod/src/generated/types/locationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */

export interface LocationInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  lat: number;
  lng: number;
  /** @minLength 1 */
  imageUrl: string;
  hint?: string;
}