# lib/api-zod/src/generated/types/locationPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */

export interface LocationPatch {
  /** @minLength 1 */
  name?: string;
  description?: string;
  lat?: number;
  lng?: number;
  imageUrl?: string;
  hint?: string;
}