# lib/api-zod/src/generated/types/locationStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */

export interface LocationStats {
  totalLocations: number;
  totalGames: number;
  avgScore: number;
}