# lib/api-zod/src/generated/types/roundResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GeoGuessr-like game API
 * OpenAPI spec version: 0.1.0
 */

export interface RoundResult {
  round: number;
  locationName: string;
  guessLat: number;
  guessLng: number;
  correctLat: number;
  correctLng: number;
  distanceKm: number;
  score: number;
}